Refactoring project.
